Sixteenth set of iterations completed

+++ cogs/report.py
@@ -1,4 +1,5 @@
 from datetime import date
+from enum import Enum
 from functools import partial
 from pathlib import Path
 
@@ -7,8 +8,9 @@ import discord
 from discord import app_commands, File, Member
 from discord.ext import commands
 
-from services.disc import requires_permission, respond
-from util.datatypes import ReportRange, SimpleReportRange
+from facades.texts import render_text
+from services.disc import member_language, requires_permission, respond
+from util.datatypes import CommandChoiceOption, ReportGranularity, ReportRange, SimpleReportRange
 from util.identifiers import PermissionFlagID, TextPieceID
 from dateutil.parser import parse as parse_datetime, ParserError
 
@@ -61,7 +63,7 @@ class ReportCog(commands.GroupCog, name="report", description="Commands for disp
 
     async def simple_report_command(
         self,
-        report_generator: tp.Callable[[SimpleReportRange], tp.Coroutine[tp.Any, tp.Any, str]],
+        report_generator: tp.Callable[[SimpleReportRange], tp.Coroutine[tp.Any, tp.Any, str | None]],
         inter: discord.Interaction,
         date_from: str | None = None,
         date_to: str | None = None
@@ -73,52 +75,62 @@ class ReportCog(commands.GroupCog, name="report", description="Commands for disp
             return
 
         image_path = await report_generator(report_range)
+        if not image_path:
+            await inter.edit_original_response(content=render_text(TextPieceID.ERROR_REPORT_NO_DATA, member_language(inter.user, inter.locale).language))
+            return
+
         await inter.edit_original_response(attachments=[File(image_path)])
         Path(image_path).unlink()
 
     async def granular_report_command(
         self,
-        report_generator: tp.Callable[[ReportRange], tp.Coroutine[tp.Any, tp.Any, str]],
+        report_generator: tp.Callable[[ReportRange], tp.Coroutine[tp.Any, tp.Any, str | None]],
         inter: discord.Interaction,
         date_from: str | None = None,
         date_to: str | None = None,
-        by_week: bool | None = None
+        granularity: ReportGranularity = ReportGranularity.DAY
     ):
         await inter.response.defer(ephemeral=True, thinking=True)
 
-        report_range = await self.prepare_range(inter, date_from, date_to, by_week)
+        report_range = await self.prepare_range(inter, date_from, date_to, granularity == ReportGranularity.WEEK)
         if not report_range:
             return
 
         image_path = await report_generator(report_range)
+        if not image_path:
+            await inter.edit_original_response(content=render_text(TextPieceID.ERROR_REPORT_NO_DATA, member_language(inter.user, inter.locale).language))
+            return
+
         await inter.edit_original_response(attachments=[File(image_path)])
         Path(image_path).unlink()
 
     @app_commands.command(description=TextPieceID.COMMAND_DESCRIPTION_REPORT_NEW_REQUESTS.as_locale_str())
     @app_commands.describe(
-        date_from=TextPieceID.COMMAND_OPTION_REPORT_NEW_REQUESTS_TS_FROM.as_locale_str(),
-        date_to=TextPieceID.COMMAND_OPTION_REPORT_NEW_REQUESTS_TS_TO.as_locale_str(),
-        by_week=TextPieceID.COMMAND_OPTION_REPORT_NEW_REQUESTS_TS_TO.as_locale_str(),  # TODO
+        date_from=TextPieceID.COMMAND_OPTION_REPORT_NEW_REQUESTS_DATE_FROM.as_locale_str(),
+        date_to=TextPieceID.COMMAND_OPTION_REPORT_NEW_REQUESTS_DATE_TO.as_locale_str(),
+        granularity=TextPieceID.COMMAND_OPTION_REPORT_NEW_REQUESTS_GRANULARITY.as_locale_str(),
     )
+    @app_commands.choices(granularity=CommandChoiceOption.report_granularity())
     @requires_permission(PermissionFlagID.REPORT_VIEWER)
-    async def new_requests(self, inter: discord.Interaction, date_from: str | None = None, date_to: str | None = None, by_week: bool = False) -> None:
-        await self.granular_report_command(facades.reports.new_requests, inter, date_from, date_to, by_week)
+    async def new_requests(self, inter: discord.Interaction, date_from: str | None = None, date_to: str | None = None, granularity: ReportGranularity = ReportGranularity.DAY) -> None:
+        await self.granular_report_command(facades.reports.new_requests, inter, date_from, date_to, granularity)
 
-    @app_commands.command(description=TextPieceID.COMMAND_DESCRIPTION_REPORT_NEW_REQUESTS.as_locale_str())  # TODO
+    @app_commands.command(description=TextPieceID.COMMAND_DESCRIPTION_REPORT_PENDING_REQUESTS.as_locale_str())
     @app_commands.describe(
-        date_from=TextPieceID.COMMAND_OPTION_REPORT_NEW_REQUESTS_TS_FROM.as_locale_str(),  # TODO
-        date_to=TextPieceID.COMMAND_OPTION_REPORT_NEW_REQUESTS_TS_TO.as_locale_str(),  # TODO
-        by_week=TextPieceID.COMMAND_OPTION_REPORT_NEW_REQUESTS_TS_TO.as_locale_str(),  # TODO
+        date_from=TextPieceID.COMMAND_OPTION_REPORT_PENDING_REQUESTS_DATE_FROM.as_locale_str(),
+        date_to=TextPieceID.COMMAND_OPTION_REPORT_PENDING_REQUESTS_DATE_TO.as_locale_str(),
+        granularity=TextPieceID.COMMAND_OPTION_REPORT_PENDING_REQUESTS_GRANULARITY.as_locale_str(),
     )
+    @app_commands.choices(granularity=CommandChoiceOption.report_granularity())
     @requires_permission(PermissionFlagID.REPORT_VIEWER)
-    async def pending_requests(self, inter: discord.Interaction, date_from: str | None = None, date_to: str | None = None, by_week: bool = False) -> None:
-        await self.granular_report_command(facades.reports.pending_requests, inter, date_from, date_to, by_week)
+    async def pending_requests(self, inter: discord.Interaction, date_from: str | None = None, date_to: str | None = None, granularity: ReportGranularity = ReportGranularity.DAY) -> None:
+        await self.granular_report_command(facades.reports.pending_requests, inter, date_from, date_to, granularity)
 
-    @app_commands.command(description=TextPieceID.COMMAND_DESCRIPTION_REPORT_NEW_REQUESTS.as_locale_str())  # TODO
+    @app_commands.command(description=TextPieceID.COMMAND_DESCRIPTION_REPORT_REVIEWER_OPINIONS.as_locale_str())
     @app_commands.describe(
-        reviewer=TextPieceID.COMMAND_OPTION_REPORT_NEW_REQUESTS_TS_FROM.as_locale_str(),  # TODO
-        date_from=TextPieceID.COMMAND_OPTION_REPORT_NEW_REQUESTS_TS_FROM.as_locale_str(),  # TODO
-        date_to=TextPieceID.COMMAND_OPTION_REPORT_NEW_REQUESTS_TS_TO.as_locale_str(),  # TODO
+        reviewer=TextPieceID.COMMAND_OPTION_REPORT_REVIEWER_OPINIONS_REVIEWER.as_locale_str(),
+        date_from=TextPieceID.COMMAND_OPTION_REPORT_REVIEWER_OPINIONS_DATE_FROM.as_locale_str(),
+        date_to=TextPieceID.COMMAND_OPTION_REPORT_REVIEWER_OPINIONS_DATE_TO.as_locale_str(),
     )
     @requires_permission(PermissionFlagID.REPORT_VIEWER)
     async def reviewer_opinions(self, inter: discord.Interaction, reviewer: Member, date_from: str | None = None, date_to: str | None = None) -> None:
@@ -129,15 +141,16 @@ class ReportCog(commands.GroupCog, name="report", description="Commands for disp
             date_to
         )
 
-    @app_commands.command(description=TextPieceID.COMMAND_DESCRIPTION_REPORT_NEW_REQUESTS.as_locale_str())  # TODO
+    @app_commands.command(description=TextPieceID.COMMAND_DESCRIPTION_REPORT_REVIEW_ACTIVITY.as_locale_str())
     @app_commands.describe(
-        date_from=TextPieceID.COMMAND_OPTION_REPORT_NEW_REQUESTS_TS_FROM.as_locale_str(),  # TODO
-        date_to=TextPieceID.COMMAND_OPTION_REPORT_NEW_REQUESTS_TS_TO.as_locale_str(),  # TODO
-        by_week=TextPieceID.COMMAND_OPTION_REPORT_NEW_REQUESTS_TS_TO.as_locale_str(),  # TODO
+        date_from=TextPieceID.COMMAND_OPTION_REPORT_REVIEW_ACTIVITY_DATE_FROM.as_locale_str(),
+        date_to=TextPieceID.COMMAND_OPTION_REPORT_REVIEW_ACTIVITY_DATE_TO.as_locale_str(),
+        granularity=TextPieceID.COMMAND_OPTION_REPORT_REVIEW_ACTIVITY_GRANULARITY.as_locale_str(),
     )
+    @app_commands.choices(granularity=CommandChoiceOption.report_granularity())
     @requires_permission(PermissionFlagID.REPORT_VIEWER)
-    async def review_activity(self, inter: discord.Interaction, date_from: str | None = None, date_to: str | None = None, by_week: bool = False) -> None:
-        await self.granular_report_command(facades.reports.review_activity, inter, date_from, date_to, by_week)
+    async def review_activity(self, inter: discord.Interaction, date_from: str | None = None, date_to: str | None = None, granularity: ReportGranularity = ReportGranularity.DAY) -> None:
+        await self.granular_report_command(facades.reports.review_activity, inter, date_from, date_to, granularity)
 
 async def setup(bot):
     await bot.add_cog(ReportCog(bot))
